requirements.txt
einops==0.8.0
kornia==0.7.2
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
pycocotools==2.0.7
skimage==0.0
torch==2.2.2
torcheval==0.0.7
torchvision==0.17.2
tqdm==4.66.4